+++ src/components/shared/ErrorMessage.jsx
@@ -2,11 +2,11 @@ import PropTypes from 'prop-types'
 import styles from './styles/ErrorMessage.module.css'
 
 const ErrorMessage = ({ error }) => (
-    <div className={styles.container}>{error.message}</div>
+    <div className={styles.container}>{error.message || error}</div>
 )
 
 ErrorMessage.propTypes = {
-    error: PropTypes.object,
+    error: PropTypes.oneOfType([PropTypes.object, PropTypes.string]).isRequired,
 }
 
 export default ErrorMessage