+++ src/utils/chart.js
@@ -9,7 +9,7 @@ export const animation = {
 
 // Date fromat YYYY-MM
 export const getYearFromId = (id) => id.substring(0, 4)
-export const getMonthFromId = (id) => id.substring(5, 7)
+export const getMonthFromId = (id) => Number(id.substring(5, 7))
 
 export const getSelectedMonths = (data, { startTime, endTime }) =>
     data.filter((d) => d.id >= startTime && d.id <= endTime)
@@ -59,4 +59,9 @@ export const vegetationCredits = {
 export const landCoverCredits = {
     href: 'https://lpdaac.usgs.gov/products/mcd12q1v061/',
     text: i18n.t('NASA LP DAAC at the USGS EROS Center / Google Earth Engine'),
+}
+
+export const elevationCredits = {
+    href: 'https://lpdaac.usgs.gov/products/srtmgl1v003/',
+    text: i18n.t('NASA / USGS / JPL-Caltech / Google Earth Engine'),
 }